move linked channel scan into forensics task, add backend throttling
